+++ pkgs/native_assets_builder/test/build_runner/build_runner_caching_test.dart
@@ -110,26 +110,35 @@ void main() async {
   test(
     'add C file, modify hook',
     timeout: longTimeout,
-    onPlatform: {
-      'linux': const Skip('https://github.com/dart-lang/native/issues/1391.'),
-    },
     () async {
       await inTempDir((tempUri) async {
         await copyTestProjects(targetUri: tempUri);
         final packageUri = tempUri.resolve('native_add/');
 
+        final logMessages = <String>[];
+        final logger = createCapturingLogger(logMessages);
+
         await runPubGet(workingDirectory: packageUri, logger: logger);
+        logMessages.clear();
         // Make sure the first compile is at least one second after the
         // package_config.json is written, otherwise dill compilation isn't
         // cached.
         await Future<void>.delayed(const Duration(seconds: 1));
 
+        final result = await build(packageUri, logger, dartExecutable);
         {
-          final result = await build(packageUri, logger, dartExecutable);
-          await expectSymbols(
-              asset: result.assets.single as NativeCodeAssetImpl,
-              symbols: ['add']);
+          final compiledHook = logMessages
+              .where((m) =>
+                  m.contains('dart compile kernel') ||
+                  m.contains('dart.exe compile kernel'))
+              .isNotEmpty;
+          expect(compiledHook, isTrue);
         }
+        logMessages.clear();
+        await expectSymbols(
+          asset: result.assets.single as NativeCodeAssetImpl,
+          symbols: ['add'],
+        );
 
         await copyTestProjects(
             sourceUri: testDataUri.resolve('native_add_add_source/'),
@@ -137,9 +146,19 @@ void main() async {
 
         {
           final result = await build(packageUri, logger, dartExecutable);
+          {
+            final compiledHook = logMessages
+                .where((m) =>
+                    m.contains('dart compile kernel') ||
+                    m.contains('dart.exe compile kernel'))
+                .isNotEmpty;
+            expect(compiledHook, isTrue);
+          }
+          logMessages.clear();
           await expectSymbols(
-              asset: result.assets.single as NativeCodeAssetImpl,
-              symbols: ['add', 'multiply']);
+            asset: result.assets.single as NativeCodeAssetImpl,
+            symbols: ['add', 'multiply'],
+          );
         }
       });
     },