+++ pkgs/swift2objc/lib/src/ast/_core/interfaces/compound_declaration.dart
@@ -2,9 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../shared/referred_type.dart';
+import '../../declarations/compounds/members/method_declaration.dart';
+import '../../declarations/compounds/members/property_declaration.dart';
 import 'declaration.dart';
-import 'parameterizable.dart';
 import 'protocol_conformable.dart';
 import 'type_parameterizable.dart';
 
@@ -13,22 +13,6 @@ import 'type_parameterizable.dart';
 /// for concrete implementations.
 abstract interface class CompoundDeclaration
     implements Declaration, TypeParameterizable, ProtocolConformable {
-  abstract List<CompoundPropertyDeclaration> properties;
-  abstract List<CompoundMethodDeclaration> methods;
-}
-
-/// An interface for a compound property. See `ClassPropertyDeclaration`,
-/// `StructPropertyDeclaration` and `ProtocolPropertyDeclaration`
-/// for concrete implementations.
-abstract interface class CompoundPropertyDeclaration implements Declaration {
-  abstract bool hasSetter;
-  abstract ReferredType type;
-}
-
-/// An interface for a compound method. See `ClassMethodDeclaration`,
-/// `StructMethodDeclaration` and `ProtocolMethodDeclaration`
-/// for concrete implementations.
-abstract interface class CompoundMethodDeclaration
-    implements Declaration, TypeParameterizable, Parameterizable {
-  abstract ReferredType? returnType;
+  abstract List<PropertyDeclaration> properties;
+  abstract List<MethodDeclaration> methods;
 }